// backend/src/entities/enums/calculator-category.enum.ts
export enum CalculatorCategory {
  Technology = 'Technology',
  Product_Development = 'Product Development',
  Product_Definition_Design = 'Product Definition Design',
  Competitive_Landscape = 'Competitive Landscape',
  Team = 'Team',
  Go_To_Market = 'Go-To-Market',
  Manufacturing_Supply_Chain = 'Manufacturing Supply Chain',
}